backend/requirements.txt
fastapi
uvicorn[standard]
redis[asyncio]
motor
pydantic
pydantic-settings
google-generativeai
websockets
python-dotenv
